requirements.txt
django==1.9.4
djangorestframework==3.3.3
django-cors-headers
markdown
pillow
psycopg2
